Centralize shorcuts menu code into one component
